requirements.txt
greenlet==3.0.3
prettytable==3.10.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
typing_extensions==4.12.2
wcwidth==0.2.13
